imageLink

+++ src/main/java/ro/itschool/repository/PaintingRepository.java
@@ -24,7 +24,7 @@ public interface PaintingRepository extends JpaRepository<Painting, UUID> {
     @Query(
             value = "SELECT * FROM painting WHERE user_id = ?",
             nativeQuery = true)
-    List<Painting> findByUserId(Integer userId);
+    List<Painting> findByUserId(Long userId);
 
     @Query(
             value = "SELECT * FROM painting p WHERE p.name LIKE %:keyword% OR p.author LIKE %:keyword% OR p.location LIKE %:keyword% OR p.movement LIKE %:keyword% " +
